Technically complete Day 18, Part 2
This program only works because Haskell is able to detect that program 0
and program 1 are in a deadlock. The program prints out whenever a
program sends or receives data so if you collect the output and pipe to
`grep '1:sending' | wc -l`, you'll get the answer. It's a huge hack.